+++ server/public/scripts/controllers/sale.controller.js
@@ -6,30 +6,35 @@ myApp.controller('SaleController', function ($http) {
     getSaleListings();
 
     function getSaleListings() {
-        console.log('in getSaleListings');   
+        console.log('in getSaleListings');
         $http({
             method: 'GET',
             url: '/sale'
-        }).then(function(response) {
+        }).then(function (response) {
             console.log('in GET-sc', response.data);
             sc.sale = response.data;
-        }).catch(function(error) {
-            console.log('error in GET-sc', error);     
+        }).catch(function (error) {
+            console.log('error in GET-sc', error);
         });
     }; //end of getSaleListings 
-    
+
     sc.removeSale = function (saleId) {
         console.log('in removeRent', saleId);
-        $http({
-            method: 'DELETE',
-            url: '/delete/' + saleId
-        }).then(function (response) {
-            alert('Listing successfully deleted!')
-            console.log('sale deleted', response);
-            getSaleListings();
-        }).catch(function (error) {
-            alert('Unable to delete!')
-            console.log('error in removeSale', error);
-        });
+        let cT = confirm('Confirming deletion?');
+        if (cT == true) {
+            $http({
+                method: 'DELETE',
+                url: '/delete/' + saleId
+            }).then(function (response) {
+                alert('Listing successfully deleted!')
+                console.log('sale deleted', response);
+                getSaleListings();
+            }).catch(function (error) {
+                alert('Unable to delete!')
+                console.log('error in removeSale', error);
+            });
+        } else {
+            alert ('Deletion cancelled.')
+        }; //end of if-else
     }; //end of removeSale
 }); //end of controller
